extract the add a new place form out as a new component and refactor the code
